backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
openai>=1.40.0
langchain==0.3.13
google-generativeai==0.8.3
langchain-google-genai==2.0.7
pinecone-client==5.0.1
httpx==0.28.1
SpeechRecognition==3.10.4
pydub==0.25.1
python-multipart==0.0.20
